test: merge ColumnLayoutState geometry tests into the existing suite, drop the duplicate file
